Update 7. equation1.c

+++ Assignment 10/7. equation1.c
@@ -23,7 +23,7 @@ float equation(int n){
         float result=1.0;
         int i;
         for(i=1;i<n;i++){
-                result+=(float)(1.0/factorial(i+1));
+                result+=(1.0/factorial(i+1));
         }
         //return result;
         printf("Result : %f \n", result);
